# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Transport System Backend API
  description: Spring Boot based back-end service for transport system
  version: V1
servers:
- url: "http://localhost:3200"
  description: Generated server url
...
paths:
...
  /api/findroute:
    post:
      tags:
      - transport system
      description: 'Find Route '
      operationId: getDirection
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Path"
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Direction"
...
components:
  schemas:
...
    Path:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
    Direction:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
        path:
          type: string
        distance:
          type: number
          format: double
        duration:
          type: number
          format: double
...